Refactor touch
